Project MVP Tkinter model/requirements.txt
Pillow==9.4.0
PyYAML==6.0
ttkthemes==3.2.2
